requirements.txt
# Sanger 16S rRNA 测序分析工具包 - 依赖项
# Python >= 3.7

# 核心生物信息学库
biopython>=1.79

# 数据处理和分析
pandas>=1.3.0
numpy>=1.21.0

# Excel 文件读写
openpyxl>=3.0.0

# 配置文件解析
PyYAML>=5.4.0

# 可选：进度条显示（推荐）
tqdm>=4.62.0

# 可选：更好的命令行参数解析
argparse>=1.4.0

